+++ tests/Gedmo/Mapping/Fixture/Document/User.php
@@ -12,11 +12,13 @@ declare(strict_types=1);
 namespace Gedmo\Tests\Mapping\Fixture\Document;
 
 use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
+use Doctrine\ODM\MongoDB\Types\Type;
 use Gedmo\Tests\Mapping\Mock\Extension\Encoder\Mapping as Ext;
 
 /**
  * @ODM\Document(collection="test_users")
  */
+#[ODM\Document(collection: 'test_users')]
 class User
 {
     /**
@@ -24,6 +26,7 @@ class User
      *
      * @ODM\Id
      */
+    #[ODM\Id]
     private $id;
 
     /**
@@ -31,6 +34,8 @@ class User
      *
      * @ODM\Field(type="string")
      */
+    #[Ext\Encode(type: 'sha1', secret: 'xxx')]
+    #[ODM\Field(type: Type::STRING)]
     private ?string $name = null;
 
     /**
@@ -38,6 +43,8 @@ class User
      *
      * @ODM\Field(type="string")
      */
+    #[Ext\Encode(type: 'md5')]
+    #[ODM\Field(type: Type::STRING)]
     private ?string $password = null;
 
     public function setName(?string $name): void