+++ tests/Gedmo/Mapping/Mock/Extension/Encoder/Mapping/Driver/Annotation.php
@@ -11,26 +11,13 @@ declare(strict_types=1);
 
 namespace Gedmo\Tests\Mapping\Mock\Extension\Encoder\Mapping\Driver;
 
-use Doctrine\Common\Annotations\AnnotationReader;
-use Doctrine\Persistence\Mapping\Driver\MappingDriver;
-use Gedmo\Mapping\Driver;
+use Gedmo\Mapping\Driver\AbstractAnnotationDriver;
 use Gedmo\Tests\Mapping\Mock\Extension\Encoder\Mapping\Encode;
 
-class Annotation implements Driver
+class Annotation extends AbstractAnnotationDriver
 {
-    /**
-     * original driver if it is available
-     *
-     * @var MappingDriver
-     */
-    protected $_originalDriver;
-
     public function readExtendedMetadata($meta, array &$config)
     {
-        $reader = new AnnotationReader();
-        // set annotation namespace and alias
-        // $reader->setAnnotationNamespaceAlias('Gedmo\Tests\Mapping\Mock\Extension\Encoder\Mapping\\', 'ext');
-
         $class = $meta->getReflectionClass();
         // check only property annotations
         foreach ($class->getProperties() as $property) {
@@ -41,8 +28,9 @@ class Annotation implements Driver
             ) {
                 continue;
             }
+
             // now lets check if property has our annotation
-            if ($encode = $reader->getPropertyAnnotation($property, Encode::class)) {
+            if ($encode = $this->reader->getPropertyAnnotation($property, Encode::class)) {
                 $field = $property->getName();
                 // check if field is mapped
                 if (!$meta->hasField($field)) {
@@ -66,13 +54,5 @@ class Annotation implements Driver
         }
 
         return $config;
-    }
-
-    /**
-     * Passes in the mapping read by original driver
-     */
-    public function setOriginalDriver($driver): void
-    {
-        $this->_originalDriver = $driver;
     }
 }