+++ CHANGELOG.md
@@ -18,9 +18,13 @@ a release.
 ---
 
 ## [Unreleased]
+### Added
+- Tree: Added `@template` and `@template-extends` annotations to the Tree repositories
+
 ### Changed
 - Dropped support for PHP < 7.4
 - Dropped support for Symfony < 5.4
+- Dropped support for doctrine/dbal < 3.2
 
 ### Deprecated
 - Calling `Gedmo\Mapping\Event\Adapter\ORM::getObjectManager()` and `getObject()` on EventArgs that do not implement `getObjectManager()` and `getObject()` (such as old EventArgs implementing `getEntityManager()` and `getEntity()`)